Add help tooltip to clarify selecting the current version of a table when a version number is required
There are cases in Synapse where the Entity Finder can be used to select a table where a version number is required (e.g. adding datasets to a dataset collection). It's possible to add tables that don't have snapshots, since you can reference the draft version, but doesn't make sense in the conceptual versioning model we try to present.